backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.11.0
PyMySQL==1.1.2
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
